Approach changed back to use 'render_block' filter due to order issues with sanitizer implementation
