+++ includes/embeds/class-amp-core-block-handler.php
@@ -6,6 +6,8 @@
  */
 
 use AmpProject\Attribute;
+use AmpProject\Extension;
+use AmpProject\Layout;
 use AmpProject\Dom\Document;
 
 /**
@@ -51,6 +53,7 @@ class AMP_Core_Block_Handler extends AMP_Base_Embed_Handler {
 		'core/archives'   => 'ampify_archives_block',
 		'core/video'      => 'ampify_video_block',
 		'core/file'       => 'ampify_file_block',
+		'core/navigation' => 'ampify_navigation_block',
 	];
 
 	/**
@@ -254,6 +257,117 @@ class AMP_Core_Block_Handler extends AMP_Base_Embed_Handler {
 	 */
 	public function dequeue_block_library_file_script() {
 		wp_dequeue_script( 'wp-block-library-file' );
+	}
+
+	/**
+	 * Ampify navigation block contained by <nav> element.
+	 *
+	 * Steps:
+	 * - "fake" the "open" state by adding `is-menu-open has-modal-open` classes to `div.wp-block-navigation__responsive-container`
+	 * - add `on="tap:navigationX.open"` to `button.wp-block-navigation__responsive-container-open` element
+	 * - add `on="tap:navigationX.close"` to `button.wp-block-navigation__responsive-container-close` element
+	 * - wrap `div.wp-block-navigation__responsive-container` with `<amp-lightbox id="navigationX" layout="nodisplay">...</amp-lightbox>`
+	 * - remove `data-micromodal-trigger` and `data-micromodal-close` attributes
+	 * - remove `aria-expanded` and `aria-modal` attributes
+	 * - duplicate `div.wp-block-navigation__responsive-container` (original one, without extra classes) outside the `amp-lightbox` wrapper and unwrap it from modal-related wrappers
+	 * - dequeue the `wp-block-navigation-view` script (see AMP_Navigation_Block_Sanitizer::add_buffering_hooks method)
+	 *
+	 * @see https://github.com/ampproject/amp-wp/issues/6319#issuecomment-978246093
+	 * @see render_block_core_navigation()
+	 *
+	 * @since 2.2
+	 *
+	 * @param string $block_content The block content about to be appended.
+	 * @param array  $block         The full block, including name and attributes.
+	 *
+	 * @return string Filtered block content.
+	 */
+	public function ampify_navigation_block( $block_content, $block ) {
+
+		$class_query = '//%1$s[ contains( concat( " ", normalize-space( @class ), " " ), " %2$s " ) ]';
+
+		$dom = new Document();
+		$dom->loadHTML( $block_content );
+
+		$node = $dom->xpath->query( '//nav' )->item( 0 );
+		if ( ! $node instanceof DOMElement ) {
+			return $block_content;
+		}
+
+		$container_node = $dom->xpath->query( sprintf( $class_query, 'div', 'wp-block-navigation__responsive-container' ), $node )->item( 0 );
+		if ( ! $container_node instanceof DOMElement ) {
+			return $block_content;
+		}
+
+		add_action( 'wp_print_scripts', [ get_class(), 'dequeue_block_navigation_view_script' ], 0 );
+		add_action( 'wp_print_footer_scripts', [ get_class(), 'dequeue_block_navigation_view_script' ], 0 );
+
+		$unique_id = $container_node->getAttribute( Attribute::ID );
+		$container_node->removeAttribute( Attribute::ID );
+
+		$cloned_container_node = $container_node->cloneNode( true );
+		$cloned_container_node->setAttribute( Attribute::CLASS_, trim( $cloned_container_node->getAttribute( Attribute::CLASS_ ) ) . ' is-menu-open has-modal-open' );
+
+		$amp_lightbox_node = AMP_DOM_Utils::create_node(
+			$dom,
+			Extension::LIGHTBOX,
+			[
+				Attribute::ID     => $unique_id,
+				Attribute::LAYOUT => Layout::NODISPLAY,
+			]
+		);
+
+		$amp_lightbox_node->appendChild( $cloned_container_node );
+		$node->appendChild( $amp_lightbox_node );
+
+		// Unwrap original container content out of "wp-block-navigation__responsive-close" and "wp-block-navigation__responsive-dialog" wrappers.
+		$content_node = $dom->xpath->query( sprintf( $class_query, 'div', 'wp-block-navigation__responsive-container-content' ), $container_node )->item( 0 );
+		$close_node   = $dom->xpath->query( sprintf( $class_query, 'div', 'wp-block-navigation__responsive-close' ), $container_node )->item( 0 );
+
+		if ( $content_node instanceof DOMElement && $close_node instanceof DOMElement ) {
+			$content_node->removeAttribute( Attribute::ID );
+
+			$container_node->appendChild( $content_node );
+			$container_node->removeChild( $close_node );
+		}
+
+		// Extend "open" and "close" buttons.
+		$open_button_node  = $dom->xpath->query( sprintf( $class_query, 'button', 'wp-block-navigation__responsive-container-open' ), $node )->item( 0 );
+		$close_button_node = $dom->xpath->query( sprintf( $class_query, 'button', 'wp-block-navigation__responsive-container-close' ), $node )->item( 0 );
+
+		if ( $open_button_node instanceof DOMElement && $close_button_node instanceof DOMElement ) {
+			$open_button_node->setAttribute( 'on', sprintf( 'tap:%s.open', $unique_id ) );
+			$close_button_node->setAttribute( 'on', sprintf( 'tap:%s.close', $unique_id ) );
+		}
+
+		// Remove unwanted attributes.
+		$unwanted_attributes = [
+			'aria-expanded',
+			'aria-modal',
+			'data-micromodal-trigger',
+			'data-micromodal-close',
+		];
+
+		foreach ( $unwanted_attributes as $unwanted_attribute ) {
+			$items = $dom->xpath->query( sprintf( '//*[@%s]', $unwanted_attribute ), $node );
+			foreach ( $items as $item ) {
+				if ( ! $item instanceof DOMElement ) {
+					continue;
+				}
+				$item->removeAttribute( $unwanted_attribute );
+			}
+		}
+
+		return $dom->saveHTML( $node );
+	}
+
+	/**
+	 * Dequeue wp-block-navigation-view script.
+	 *
+	 * @since 2.2
+	 */
+	public static function dequeue_block_navigation_view_script() {
+		wp_dequeue_script( 'wp-block-navigation-view' );
 	}
 
 	/**
